+++ debian-live/build.sh
@@ -50,7 +50,7 @@ function createBuildDir {
 function configImage {
   loginfo "${FUNCNAME[0]}" "Configuring the Debian image"
 
-  local liveConfigOptions=""
+  local liveConfigOptions="" bootsplash
   if [ "${DEBIAN_SUDO_DISABLE}" == "true" ]; then
     liveConfigOptions+=" noroot"
   fi
@@ -60,6 +60,9 @@ function configImage {
   else
     liveConfigOptions+=" nopersistence"
   fi
+
+  # for the (plymouth bases) boot splash images
+  bootsplash="nvidia-drm.modeset=1 radeon.modeset=1 amdgpu.modeset=1 i915.modeset=1 silent quiet splash" 
 
   cd "${BUILD_DIR}"
   lb config \
@@ -81,7 +84,7 @@ function configImage {
     --chroot-squashfs-compression-type "${DEBIAN_SQUASHFS_COMPRESSION_TYPE}" \
     --compression xz \
     --archive-areas "main non-free-firmware" \
-    --bootappend-live "boot=live config hostname=iksdp-${RELEASE_VERSION} locales=${DEBIAN_LOCALES} keyboard-layouts=${DEBIAN_KEYBOARD_LAYOUTS} timezone=${DEBIAN_TIMEZONE} ${liveConfigOptions}" \
+    --bootappend-live "boot=live config hostname=iksdp-${RELEASE_VERSION} locales=${DEBIAN_LOCALES} keyboard-layouts=${DEBIAN_KEYBOARD_LAYOUTS} timezone=${DEBIAN_TIMEZONE} ${bootsplash} ${liveConfigOptions}" \
     --image-name debian-live-"${DEBIAN_VERSION}"-"${RELEASE_VERSION}"-"${IMAGE_TIMESTAMP}"
   if [ "$?" -ne 0 ]; then
     logerror "${FUNCNAME[0]}" "Debian image configuration failed"